requirements.txt
# Auth
msal>=1.31.0

# HTTP
requests>=2.32.0
httpx>=0.27.0

# LLM
anthropic>=0.40.0

# Embeddings & ranking
numpy>=2.0.0
scikit-learn>=1.5.0

# Storage
# sqlite3 is stdlib — no install needed

# UI
streamlit>=1.41.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
rich>=13.9.0
